+++ fsspec/asyn.py
@@ -4,7 +4,9 @@ import functools
 import inspect
 import os
 import re
+import sys
 import threading
+from contextlib import contextmanager
 from glob import has_magic
 
 from .exceptions import FSTimeoutError
@@ -87,6 +89,22 @@ def sync_wrapper(func, obj=None):
     return wrapper
 
 
+@contextmanager
+def _selector_policy():
+    original_policy = asyncio.get_event_loop_policy()
+    try:
+        if (
+            sys.version_info >= (3, 8)
+            and os.name == "nt"
+            and hasattr(asyncio, "WindowsSelectorEventLoopPolicy")
+        ):
+            asyncio.set_event_loop_policy(asyncio.WindowsSelectorEventLoopPolicy())
+
+        yield
+    finally:
+        asyncio.set_event_loop_policy(original_policy)
+
+
 def get_loop():
     """Create or return the default fsspec IO loop
 
@@ -97,7 +115,8 @@ def get_loop():
             # repeat the check just in case the loop got filled between the
             # previous two calls from another thread
             if loop[0] is None:
-                loop[0] = asyncio.new_event_loop()
+                with _selector_policy():
+                    loop[0] = asyncio.new_event_loop()
                 th = threading.Thread(target=loop[0].run_forever, name="fsspecIO")
                 th.daemon = True
                 th.start()